# refactor(mithril-client): Optimize cardano db artifacts download