--- a/mithril-client/Cargo.toml
+++ b/mithril-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mithril-client"
-version = "0.11.13"
+version = "0.11.14"
 description = "Mithril client library"
 authors = { workspace = true }
 edition = { workspace = true }