--- a/mithril-client/src/utils/mod.rs
+++ b/mithril-client/src/utils/mod.rs
@@ -4,9 +4,9 @@
 cfg_fs! {
     mod stream_reader;
     mod fs;
-    mod vec_extensions;
+    mod vec_deque_extensions;
 
     pub use stream_reader::*;
     pub use fs::*;
-    pub use vec_extensions::VecExtensions;
+    pub use vec_deque_extensions::VecDequeExtensions;
 }