## Content

This PR refactor the way cardano database v2 files are downloaded in the `mithril-client`:

- Unified download system for immutables in ancillary:
  - allowing them to be in the same join set
  - download tasks are now precomputed
- Fail early: as soon as the last location fails when downloading a files (instead of failing at the very end of the process)
- Continuously download `max_parallel_downloads` files:  No more threshold effect because of chunking

### Performance improvement over a download of ~10 000 immutables without ancillary

> [!NOTE]
> Those commands are run on a 1 Gbit/s download speed internet connection

Before, 10k immutables download took ~33 seconds:
```shell
$./mithril-client --unstable --run-mode dev cardano-db-v2 download latest --start 7345
1/6 - Checking local disk info…
2/6 - Fetching the certificate and verifying the certificate chain…
  Certificate chain validated
3/6 - Downloading and unpacking the cardano db snapshot
   [00:00:33] [############################################################################################] Files: 10,006/10,006 (0.0s)
4/6 - Computing and verifying the Merkle proof…
5/6 - Computing the cardano db snapshot message
6/6 - Verifying the cardano db signature…
Cardano database snapshot '5845e9e1232f91c804e483eda1e7af6b9816174d0410923da6f032c1c1c91f5f' archives have been successfully unpacked.
``` 

After, 10k immutables download took ~19 seconds:
```shell
$./mithril-client --unstable --run-mode dev cardano-db-v2 download latest --start 7345
1/6 - Checking local disk info…
2/6 - Fetching the certificate and verifying the certificate chain… 
  Certificate chain validated
3/6 - Downloading and unpacking the cardano db snapshot
   [00:00:19] [############################################################################################] Files: 10,006/10,006 (0.0s)
4/6 - Computing and verifying the Merkle proof…
5/6 - Computing the cardano db snapshot message
6/6 - Verifying the cardano db signature…
Cardano database snapshot '5845e9e1232f91c804e483eda1e7af6b9816174d0410923da6f032c1c1c91f5f' archives have been successfully unpacked.
``` 

## Pre-submit checklist

- Branch
  - [x] Tests are provided (if possible)
  - [x] Crates versions are updated (if relevant)
  - [ ] CHANGELOG file is updated (if relevant)
  - [x] Commit sequence broadly makes sense
  - [x] Key commits have useful messages
- PR
  - [x] All check jobs of the CI have succeeded
  - [x] Self-reviewed the diff
  - [x] Useful pull request description
  - [x] Reviewer requested
- Documentation
  - [x] No new TODOs introduced

## Comments

This PR also fix the clippy warnings that were raised when checking with only the `fs` feature enabled.

## Issue(s)

Closes #2327